Facilitate checking of interface compatibility
In order to define the compatiblity between classes, a property restriction is
is added for the relation that connects compatible classes (here: compatibleWith).

This allow to add a lazy extension of the knowledge base for defining compatiblity
between individual under the assumption that compatibility will be defined
on class/concept level.